object-detection-device/ObjectDetectionDeviceEdgeSolution/modules/ObjectDetectionDeviceModule/requirements.txt
azure-iot-device~=2.0.0
aiohttp==3.7.2
azure-storage-queue==12.1.5